[frontend] move InteractiveLineReader to reader
